+++ crates/facade/src/use_case.rs
@@ -1,8 +1,11 @@
+use std::io;
+
 use qsimplify::{
     DiracFormat, PiFormat,
     dto::{GateOperation, mapper},
     simplifier,
 };
+use qsimplify_presenter::GraphvizFormat;
 
 #[must_use]
 pub fn display_graph(operations: &[GateOperation]) -> String {
@@ -20,6 +23,14 @@ pub fn display_grid(operations: &[GateOperation]) -> String {
 pub fn display_matrix(operations: &[GateOperation]) -> String {
     let graph = mapper::operations_to_graph(operations);
     graph.display_matrix(DiracFormat::Fancy)
+}
+
+pub fn to_graphviz(
+    operations: &[GateOperation],
+    format: GraphvizFormat,
+) -> Result<Vec<u8>, io::Error> {
+    let graph = mapper::operations_to_graph(operations);
+    qsimplify_presenter::graph_to_graphviz(&graph, format, None)
 }
 
 #[must_use]