+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - `library`: Display a quantum graph as a list of nodes and edges or as a less detailed 2D matrix.
 - `library`: Calculate the unitary matrix of a quantum graph.
 - `library`: Display a quantum graph as a unitary matrix with Dirac bra-ket notation.
+- `presenter`: Allow saving a quantum graph as a graphviz graph in GV, PNG and SVG formats.
 - `facade`: Create a DTO that stores a list of quantum gate operations in insertion order.
 - `facade`: Expose a use case for displaying the provided circuit as a graph.
 - `qiskit`: Create bindings that receive a Qiskit `QuantumCircuit` as input.